Ajout de n octet dans basebackup
